Create the pokemon_list outside of the Database
Do all the data loading and create all the Pokemon BEFORE creating the Database.  This will greatly streamline the Database class by allowing it to focus on runtime manipulation of data instead of the complexities of data loading.  This approach creates namedtuples instead of classes for both Pokemon and Region.  NamedTuples are fast, lightweight, read-only and self documenting.

The paths are too long, we are storing too much data.  To save space, they should be relative paths from IMAGE_DIR.