Move list, dict from class level to instance level
Creating two instance of the Database class throws `Exception: Duplicate names detected.`

You can see this by running the code:
```python
db0 = Database()  # create the fist db... no problems
db1 = Database()  # Exception: Duplicate names detected
```

This is caused because all instances of a class share each class level variable so when db1 wants to add a pokemon to the list, it sees that db0 has already added it.  Class-level variables are good for constants but in general should be avoided for things that get modified at runtime unless shared state between all instances is desireable.

This is a problem because pytest will create multiple databases as it sets up, executes, and tears down tests various aspects of the app.  The cleanest fix to this problem that I can think of is to demote the list and dict from being class level to being instance level.  This way each instance has its own separate copy and we reduce shared state